+++ src/DataSources/CSharp/Chunker/ICSharpChunker.cs
@@ -12,6 +12,7 @@ public interface ICSharpChunker
     /// Parses the provided code and returns the discovered code entities.
     /// </summary>
     /// <param name="code">The code to analyze.</param>
+    /// <param name="options">Options for the chunking of C#</param>
     /// <returns>A list of discovered code chunks.</returns>
-    List<CSharpChunk> GetChunks(string code);
+    List<CSharpChunk> GetChunks(string code, CSharpChunkerOptions? options = null);
 }